model repository service

+++ src/main/java/com/example/librarySystem/domain/repository/LendLogRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.librarySystem.domain.model.LendLog;
 
-public interface LendLogRepository extends JpaRepository<LendLog, Integer> {
+public interface LendLogRepository extends JpaRepository<LendLog, Long> {
 	
 	List<LendLog> findByUserId(String userId);
