+++ src/main/java/com/example/librarySystem/domain/repository/LendingRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.librarySystem.domain.model.Lending;
 
-public interface LendingRepository extends JpaRepository<Lending, Integer> {
+public interface LendingRepository extends JpaRepository<Lending, Long> {
 	
 	List<Lending> findByUserId(String userId);
